pyproject.toml
[project]
name = "caiv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]
